Arreglos

+++ server/controllers/productosController.js
@@ -61,7 +61,7 @@ module.exports.create = async (request, response, next) => {
     data: {
       nombre: producto.nombre,
       descripcion: producto.descripcion,
-      precio: parseInt(producto.precio),
+      precio: parseFloat(producto.precio),
       cantidad: parseInt(producto.cantidad),
       categoriaId: parseInt(producto.categoria),
       estadoProductoId: parseInt(producto.estado),
